+++ app/models/guild.rb
@@ -1,11 +1,9 @@
 class Guild < ApplicationRecord
-  has_many :users,          dependent: :nullify
+  has_many :users,          dependent: :destroy
   has_many :roles,          dependent: :destroy
   has_many :squads,         dependent: :destroy
   has_many :missions,       dependent: :destroy
   has_many :events,         dependent: :destroy
-  has_many :achievements,   dependent: :destroy
-  has_many :certificates,   dependent: :destroy
 
   validates :name,
             presence: true,