+++ app/models/role.rb
@@ -4,10 +4,6 @@ class Role < ApplicationRecord
   has_many :user_roles, dependent: :destroy
   has_many :users, through: :user_roles
 
-  has_many :role_certificate_requirements, dependent: :destroy
-  has_many :required_certificates, through: :role_certificate_requirements,
-                                   source: :certificate
-
   validates :name,
             presence: true,
             length: { maximum: 50 }