+++ db/migrate/20260113158934_create_audit_logs.rb
@@ -1,8 +1,8 @@
 class CreateAuditLogs < ActiveRecord::Migration[8.1]
   def change
     create_table :audit_logs do |t|
-      t.references :user, null: true, foreign_key: true
-      t.references :guild, null: true, foreign_key: true
+      t.references :user, null: true, foreign_key: true, on_delete: :nullify
+      t.references :guild, null: true, foreign_key: true, on_delete: :nullify
 
       t.string :action, null: false